ams-backend/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
motor==3.3.1
python-dotenv==1.0.1
// use this" pip install -r requirements.txt"